backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.2.0
django-extensions
